requirements.txt
requests~=2.28.2
pip~=22.3.1
wheel~=0.38.4
setuptools~=65.5.1
certifi~=2022.12.7
idna~=3.4
urllib3~=1.26.15
Pillow~=9.5.0